Changes to deal with memory access error in solaris.  At least I hope. Also change to description at request of CRAN maintainers

+++ src/AlleleObj.h
@@ -133,7 +133,7 @@ public:
   void mutate ();
 
   ///returns length of sequence
-  int SeqLen () ;                       
+  size_t SeqLen () ;                       
  
 
   inline void SetState (SequenceType seq)
